move site config type definitions to types.ts
