+++ src/data/site-config.ts
@@ -1,42 +1,4 @@
-export type Image = {
-    src: string;
-    alt?: string;
-};
-
-export type Link = {
-    href: string;
-    text: string;
-};
-
-export type SocialLink = Link & {
-    icon: 'bluesky' | 'codepen' | 'dev' | 'facebook' | 'github' | 'instagram' | 'linkedin' | 'mastodon' | 'medium' | 'reddit' | 'x' | 'youtube';
-};
-
-export type Hero = {
-    title?: string;
-    text?: string;
-    avatar?: Image;
-    backgroundImage?: Image;
-};
-
-export type Subscribe = {
-    title?: string;
-    text?: string;
-    formUrl: string;
-};
-
-export type SiteConfig = {
-    logo?: Image;
-    title: string;
-    description: string;
-    image?: Image;
-    primaryNavLinks?: Link[];
-    secondaryNavLinks?: Link[];
-    socialLinks?: SocialLink[];
-    hero?: Hero;
-    subscribe?: Subscribe;
-    postsPerPage?: number;
-};
+import type { SiteConfig } from '../types';
 
 const siteConfig: SiteConfig = {
     logo: {
